pyproject.toml
[tool.poetry]
name = "diary"
version = "1.0"
description = "This my own LLM app for a personal digital diary"
authors = ["Aamir Ahmad Ansari <zaikali22@gmail.com>"]
license = "MIT"
readme = "README.md"

packages = [
    { include = "diary", from = "." },
]

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.29.0"
pandas = "^2.1.4"
transformers = "^4.36.2"
torch = "1.11"
sqlalchemy = "^2.0.24"
einops = "^0.7.0"
accelerate = "^0.25.0"
fastapi = {extras = ["all"], version = "^0.108.0"}
google-generativeai = "^0.3.2"
uvicorn = {extras = ["standard"], version = "^0.25.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
